third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/sandbox-allow-scripts-subframe-expected.txt

CONSOLE MESSAGE: line 1: PASS
This test passes if it does alert pass. 
